+++ src/core/constants.js
@@ -4,7 +4,7 @@ import * as os from 'os';
 const CONSTANTS = {
     NEXO_VERSION: "1.0.3 devel",
     NEXO_VERSION_DEVEL: "n2.2.5-1",
-    NEXO_CODE: "Beta:n:2.2.5-1:AR,NOI",
+    NEXO_CODE: "Beta:n:2.2.5-1:FED,NOI",
     NODE_CORE_VERSION: process.versions.node,
     NODE_V8_VERSION: process.versions.v8,
     NODE_CPP_VERSION: process.versions.modules,