+++ src/lib/closures.js
@@ -1,19 +1,8 @@
-/**
- * 
- * @param {Post} a 
- * @param {Post} b 
- * @returns compare
- */
-const PostSortingFunction = (a,b) =>{
-    return a.datz.compareWith(b.datz)?1:-1;
-}
-
 const Nil = () => {}
 
 const notFake = (v) => v!=''
 
 export {
-    PostSortingFunction,
     Nil,
     notFake
 }